Add debug output

+++ src/test/java/kieker/monitoring/probe/aspectj/operationExecution/Util.java
@@ -36,12 +36,17 @@ public enum Util {
       } else {
          LOG.info("Deleting old log folder: " + logFolder.getAbsolutePath());
          for (File kiekerLogFolder : logFolder.listFiles()) {
-            for (File containedFile : kiekerLogFolder.listFiles()) {
-               LOG.info("Inner log file: " + containedFile);
+            if (kiekerLogFolder != null) {
+               for (File containedFile : kiekerLogFolder.listFiles()) {
+                  LOG.info("Inner log file: " + containedFile);
+               }
+               LOG.info("Kieker log folder: " + kiekerLogFolder.getAbsolutePath());
+               boolean success = kiekerLogFolder.delete();
+               LOG.info("Deleted: " + success + " Exists: " + kiekerLogFolder.exists());
+            }else {
+               LOG.error("listFiles returned null value: " + kiekerLogFolder);
             }
-            LOG.info("Kieker log folder: " + kiekerLogFolder.getAbsolutePath());
-            boolean success = kiekerLogFolder.delete();
-            LOG.info("Deleted: " + success + " Exists: " + kiekerLogFolder.exists());
+            
          }
          FileUtils.cleanDirectory(logFolder);
          LOG.info("Log folder cleaned: " + logFolder.exists());
